move to correct package DelegationTokenManager
